+++ cynic-codegen/src/inline_fragments_derive/inline_fragments_impl.rs
@@ -1,9 +1,18 @@
 use proc_macro2::TokenStream;
+use quote::quote_spanned;
+use syn::spanned::Spanned;
+
+#[derive(Clone)]
+pub enum Fallback {
+    UnionUnitVariant(syn::Ident),
+    UnionVariantWithTypename(syn::Ident, syn::Type),
+    InterfaceVariant(syn::Ident, syn::Type),
+}
 
 pub struct InlineFragmentsImpl<'a> {
     pub(super) target_enum: syn::Ident,
     pub(super) fragments: &'a [super::Fragment],
-    pub(super) fallback: Option<(syn::Ident, Option<syn::Type>)>,
+    pub(super) fallback: Option<Fallback>,
 }
 
 impl quote::ToTokens for InlineFragmentsImpl<'_> {
@@ -19,10 +28,18 @@ impl quote::ToTokens for InlineFragmentsImpl<'_> {
             .collect::<Vec<_>>();
 
         let fallback = match &self.fallback {
-            Some((variant, None)) => quote! {
+            Some(Fallback::UnionUnitVariant(variant)) => quote! {
                 Ok(#target_enum::#variant)
             },
-            Some((variant, Some(ty))) => quote! {
+            Some(Fallback::UnionVariantWithTypename(variant, ty)) => {
+                let ty_span = ty.span();
+                quote_spanned! { ty_span => {
+                        ::cynic::assert_type_eq_all!(#ty, String);
+                        Ok(#target_enum::#variant(typename.to_string()))
+                    }
+                }
+            }
+            Some(Fallback::InterfaceVariant(variant, ty)) => quote! {
                 <#ty as ::cynic::serde::Deserialize<'de>>::deserialize(deserializer).map(
                     #target_enum::#variant
                 )