+++ tests/ui-tests/tests/cases/inline-fragment-fallback-validation.rs
@@ -12,6 +12,7 @@ mod schema {
 enum MyFailingUnionType {
     Nested(Nested),
 
+    // This is allowed, but needs to be a string to work.
     #[cynic(fallback)]
     Other(Nested),
 }
@@ -33,4 +34,16 @@ enum MyOkUnionTYpe {
 
     #[cynic(fallback)]
     Other,
+}
+
+#[derive(cynic::InlineFragments)]
+#[cynic(
+    schema_path = "../../../cynic/src/bin/simple.graphql",
+    graphql_type = "MyUnionType"
+)]
+enum UnionTypeWithStringFallback {
+    Nested(Nested),
+
+    #[cynic(fallback)]
+    Other(String),
 }